requirements.txt
spaces
gradio
accelerate
iso639-lang
librosa
git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git
